pyproject.toml
[project]
name = "traderz"
version = "0.1.0"
description = "AI-powered paper trading with technical analysis and Alpaca execution"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.43.2",
    "pandas>=3.0.1",
    "ta>=0.11.0",
    "yfinance>=1.2.0",
]
